Fix inventory creation logic
it wasn't working.
It didn't make sense to have a zero paramenters constructor, we
maintain the scheme we choese for the other elements: "with id" for
retrieving the entity and a "no id" for creating the entity. the new
implementation uses the no id way as do the other impl for creating
the db entities.