Update controllers return type to int
we need to retrieve the invenotry and personalData ids to generate
the inventory entity in the db, because we can only do this after
we've created these two.
There were a couple of directions, all with their pros and cons,
this one is fine but suboptimal because we have to also handle when
we could get an error, but we also have a bigger problem.
Currently if one of the create method fails it doesn't undo the db
update, meaning we actually run the risk of generating an invenotry
and personalData but have no person that goes with it, a gap in the
db. we could fix this in many ways, but for a demo we shall make it
seem ok.
[todo]
createUser should also be a int type but because we are also doing
checks inside this means that too needs to change, i don't like it
at all, but for sake of it making sense it shall be like that.