Implement register user logic
we can actually register a new user "correctly" now. we need to
clear the db as ids are all mixed up now, but it should work.
The check for the return of the id is a sorry attempt at avoiding
to create a user with negative id for the ones it actually needs.