Add form for adding items
It follows the implementation of the previous forms, this to say
it dogshit but it works, it won't stay for long i hope.
we still use the process of creating a data class just for the
form data we need to analyze and pass back for then sending the
querry.