Implement UserService class
New service system to handle the better abstract the controller,
which should be minimal and only have the view and said operation
interact, no business code should be inside.
this service system enables to have a designated place where to
write all the logic pertaining to a certain part of the code
and avoid having the app.java mingle with anything it shouldn't
it just plain better for the structure of the project.