Update app class to just initialization
Following the new code structure, the app class is the entry point
of the application, no logic is done here just creating
the classes that will be used in the app.
1- We start by creating the frame and the manager for the view and its
pannels
2- We create the services and the associeted DAOs for logic handling
3- We Provide the controllers with the initialized services to handle
the business logic and the view manager to operate the changes to the
view as the service responds.
4- We finish by creating the appController, which wires all the
views outputs to the corresponding services providing the event
listeners for each interaction and we set the frame to be visible.