Add AppController to handle interaction view / controllers
The appcontroller is responsible for setting the event listeners
and wiring the output of the pannels to the corresponding
service + method combination, this way it only handles the wiring
between what is visible and when to the actual operation without
having any business logic inside.
Just add the initialized controller to the constructor
the pannel and its listeners and set the operation inside, finish
by adding the corresponding pannel to the viewmanager with a
suitable name.

at the bottom the first page that is show at the app launch is setted
it can be changed for debuggin purposes.