Add SignInPanel
new pannel system, each panel is responsible for the gui and what is shown
to the user.
This pannel is responsible for displaying on the frame the Welcome page or
Sign-In page where a user can either log in if already registered or create
a new account to use the app.
The system relies on a modular component system that can be reused as needed;
in this case we use the styledButton componenent for creating buttons with
a label and an icon, more components can be added, like the header for
displaying a title and the likes.
Each pannel works by implementing a createPanel method, which is called
by the constructor, here we specify the pannel components and how it
should look inside of a JPanel object.
We then create an ActionListener and its relative setter and we tigh
everything together with a ActionListener interface which implements
the methods that the AppController calls when we execute a certain
action like pressing a button. The listeners are added inside the
create buttonPanel, a helper method or class could be implemented
later on to simplify the code further but we don't really have to
manage to many things for now.