Add LogInPanel
follow the logic described in the previous commit for the SignInPanel.
It needs to be updated to have a cleaner look overall but for now it
implements everything that is needed for a log in section.
Here is an example of a action listner that takes a certain input
from some text fields which are then routed to the correct service
via the AppController and sub-entityController which manages said
service.