Add ViewManager
The ViewManager is responsible for, as the name implies, managing
the different Panels, it simply provided a way to register, show and
get a specified pannel via a provied name.
When registering a panle we map it to a provied name and the corresponding
panel we've created separately.
To other operations simply just need to retrieve a already created
pannel via its name, making it extreamly easy to work with.
The viewManager is created on a provied fram, currently when we
initialize the app we pass it the one and only main frame, but
some logic for other specific frames could be implemented, like
some special situation that requires a different frame to be drawn