Update the UserController
Reworked logic and functions of the controlle to match the new architecture,
we implement just the methods that handle operations pertaining
to a user, like creating a new user or authenticate an existing user.
The controller should be resposible for only wiring the view and the
service togheter, without implementing any business logic.
For now the only implemented method is the handleAuth
which simply gets the result from the service and responds by changing
the shown pannel via the viewManager.
all the other controllers will have to be written this way.