Remove PersonalDataController
Because of the new architecture the controller is just responsible
of wiring the view with the business logic inside the service.
the controller is needed for those "entities" with which we interact
direcltly, like the user or the inventory, but we never actually
interact with the personal data if not when we create the user.
Because of these reasons we don't need either a controller nor a
service for this entity - will add in future if needed