Update UserDAO and UserService to handle setting the session user
We get the DAOresult of the registration / log in operation and because
it contains the UserId for said Operation we simply return the user
that matches that unique id and set it as the current session user.

this could have been handled in the DAO or the service but because
of the logic it implements I dediced to implement it here, at the cost
of having a DAO method that returns a user Instead of a DAOResult
which should always be used for DAO operations.
this way the logic is all here and can be easly read.