Add HomePanel
this is the main pannel that we see as soon as we enter the app,
it consists of a top bar with the user info and two sections below
the view of All the inventories in the platform and the buttons which
handle the main operations.
when creating this pannew we also have to have created its subpannel
and pass them to it for this to work, no listeners need to be passed
we will handle them in the controller.