Implement structure for home panel in AppController
for now just the left side buttons are defined and working
but the ground work for the other pannels is laid out.
listners setup is done just add the methods once they are created
inside their respective pannels.
We also create and register the home pannel, which is conposed by
subpannels and doesn't need any listeners by itself, pretty nifty :)