Update inventory panel logic to use current user id
we call the session manager to get the current session user and
set generate or refresh the pannel based on if the pannel exists.

IMPORTANT - we create the event listeners like the exit button and
handle them in this separate function instead of the main body
because we need to have a valid istance of the pannel, aka a
generate one and that case occurs only if the pannel is null meaning
it doesn't exist so we create one and add setup the listeners.
this shouldn't interfere with the logic we have for the refresh button
as in that case we simply get once again the items inside but
the basic pannel is done