Update appController to only handle static pannels creation
will have to create a viewManager or service to handle the
update of a pannel once its created, but as of now its mutch more
important to get everything working. handling dynamic pannels that
refresh and update based on the user is too mutch of an hastle as
of now, ideally we would draw the static elements always, take
an output of a service that returns the model elements that contain
the infos to be displayed, update the already created staic elements
and then move forward.