Add inventoryController method for updates
This defines the structure of service --> view method on a controller.
Should be referenced in the future for implementing similar behaviour.
The pannel is statically created in the AppController as a basic no data
pannel. We then call a Update method for handling the generating the data
based on the service result, say the one connected to the current usser
which is available only after the log in or registration; but also for other
cases this can be referenced. As a result we have a system that can handle
dynamically changing pannels without te need of further abstraction.