Update to new viewDataPayload system + titles
we now can display the proper title to on the view of the items
in the inventory.
Thanks to this we finally are "done" with this view, we had to also
pass the mediaDAO to the service in order to handle the retrieval
of all items titles.
We now have a "system" for passing the data to the view
that is general enough that it can be used anywhere we may need it.