Add home panel manager
A new manager has been haded for handling the creation of the home
panel. Because the panel is a composition of subpanels, some of which
are dynaminc, meaning they can only be generated after a log in, we
define a manager that handles this behaviour.
The idea here is that in the userController after a succesfull log
in we define and method that hanles the update and so the setting
of the desired listeners, by doing so it will call the manager which
at creation already has some pannels provied, those that are static
simply stay in the appController, and those that are dynamic can be
set at will and the by simply calling the manager update method the
new pannel version is registered.

TODO
- Generalize this manager to hanle all the dynamic pannels so
that it is more suitable and not use case dependant.
- Maybe this way of doing the controlers is wrong, research.