Fix bug for in view for empty inv
if the inventory was empty, say a user just regisred, the logic
wasn't working properly for atleast displaying  the exit button.
We simply chek if the view data payload is null, if so we skip
parsing the data and because all the sublists are null by default
we already had the logic to handle that, so it works now.