Update to support users inv
moved the logic here, before it was initialized by the view manager
but never updated at login, causing the data to be loaded at startup
instead of after the login. The logic is still handled in the
panel but could be moved later on.
similar to other handle methods for the update we create the event
listener for the updated panel and pass it to the dynamic panel manager
for handling the updates