Fix working logic for trade panel
based on the inv-showcase provided code we fixed the previous impl
by adjusting it to the new returns.
its basically the same as before + some bug fixes that did't allow
the code to operate as expected. we get passed the user to target
for the trade request, make a service call to get its inv+titles and
pass it to the view, there its simply displayed and we get a list with
all the current and target user items involved in the trade.