Update trade service
we now also allow for the list to be null, but we will always get
atleast one item as its handled by the view, change it here if need be.
we also register the item as item in trade now.

todo
check if the flip is right in the second registration loop, where the
wanted items are registerd as in trade